[fix]스탯 생성 및 스탯 리스트 조회

+++ src/main/java/site/FitUp/main/api/stat/dtos/StatResponse.java
@@ -27,7 +27,7 @@ public class StatResponse {
     @Builder
     @Getter
     public static class GetStatResponse{
-        private int userStatSeq;
+        private long userStatSeq;
         private int strength;
         private int endurance;
         private int speed;
